requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
pep8==1.7.1
numpy==1.21.0
pytest==6.2.4
resource==0.2.1
flake8==4.0.1
torch==1.9.0+cpu
torchvision==0.10.0+cpu
torchaudio==0.9.0
pytorch-lightning==1.5.10
Shapely==1.8.1.post1
geopandas==0.10.2
pandas==1.4.1
boto3==1.21.26
botocore==1.24.26
rasterio==1.2.10
PyYAML==6.0